// src/features/editableProfileCard/model/consts/consts.ts
export enum ValidateProfileErrors {
    NO_DATA = 'NO_DATA',
    SERVER_ERROR = 'SERVER_ERROR',
    INCORRECT_NAMES = 'INCORRECT_NAMES',
    INCORRECT_AGE = 'INCORRECT_AGE',
    INCORRECT_COUNTRY = 'INCORRECT_COUNTRY',
    INCORRECT_CITY = 'INCORRECT_CITY',
    INCORRECT_CURRENCY = 'INCORRECT_CURRENCY',
    INCORRECT_AVATAR = 'INCORRECT_AVATAR',
}